pull optional component utils into separate dir
