+++ lib/component-utils/controlled-component.js
@@ -1,5 +1,5 @@
-import Component from './component';
-import {EMPTY_DIV} from './dom-patcher';
+import Component from '../component';
+import {EMPTY_DIV} from '../dom-patcher';
 
 export default class ControlledComponent extends Component {
   constructor() {
