+++ lib/component-utils/proxy-component.js
@@ -1,5 +1,6 @@
-import {h} from 'snabbdom';
-import Component from './component';
+import Component from '../component';
+import {h} from '../dom-patcher';
+
 
 /**
  * Definition of a generic Component which proxies attributes and events
