+++ lib/index.js
@@ -9,15 +9,22 @@
  */
 
 import Component from './component';
-import ControlledComponent from './controlled-component';
-import ProxyComponent from './proxy-component';
-import StateController from './state-controller';
-import StateStore from './state-store';
+import ComponentUtils from './component-utils';
 import {h} from './dom-patcher';
+
+const {
+  ControlledComponent,
+  ProxyComponent,
+  StateController,
+  StateStore,
+} = ComponentUtils;
 
 export {
   /** {@link Component} class, to be subclassed by apps */
   Component,
+
+  /** {@link component-utils} wrappers and utilities */
+  ComponentUtils,
 
   /** {@link ControlledComponent} class, to be subclassed by apps */
   ControlledComponent,
